Added Polestar Website files
Changed index.html to apache-config.html: this is so index no longer
loads when going straight on the site.

All files are in the root folder and the DB user/pass has not been
changed.

In theory this should just work on your machine as all of the code is
exactly as you sent it. (untouched.. just copied and pasted)

The file in:
wp_content>themes>mh-purity-child>page-booking.php

is the linking file that formats the booking page (clicking the link on
the navbarmenu)

To get this to work you should point your apache2 web server at it, or
copy and paste into web directory.